AI_SAFETY_PRJ_1-main/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
python-multipart>=0.0.9,<1.0
opencv-python>=4.10,<5.0
mediapipe>=0.10.30,<1.0
numpy>=1.26,<3.0
httpx>=0.27,<1.0
pydantic>=2.8,<3.0
python-dotenv>=1.0,<2.0
tzdata>=2024.1,<2027.0
pytest>=8.0,<9.0
